Editable에서 상하단 바를 고정시키는 코드를 다른 컴포넌트로 분리
